DOP-2367 - Refactor various QuizWidget pieces of code
