+++ src/components/ComponentFactory.js
@@ -43,8 +43,8 @@ import OpenAPI from './OpenAPI';
 import Operation from './Operation';
 import Paragraph from './Paragraph';
 import Procedure from './Procedure';
-import QuizChoice from './QuizChoice';
-import QuizWidget from './QuizWidget';
+import QuizChoice from './Widgets/QuizWidget/QuizChoice';
+import QuizWidget from './Widgets/QuizWidget/QuizWidget';
 import Reference from './Reference';
 import RefRole from './RefRole';
 import ReleaseSpecification from './ReleaseSpecification';
