+++ src/components/Widgets/QuizWidget/QuizWidget.js
@@ -5,10 +5,10 @@ import Button from '@leafygreen-ui/button';
 import Card from '@leafygreen-ui/card';
 import { uiColors } from '@leafygreen-ui/palette';
 import Icon from '@leafygreen-ui/icon';
-import ComponentFactory from './ComponentFactory';
-import { theme } from '../theme/docsTheme';
-import { useSiteMetadata } from '../hooks/use-site-metadata';
-import { getPlaintext } from '../utils/get-plaintext';
+import ComponentFactory from '../../ComponentFactory';
+import { theme } from '../../../theme/docsTheme';
+import { useSiteMetadata } from '../../../hooks/use-site-metadata';
+import { getPlaintext } from '../../../utils/get-plaintext';
 
 const StyledCard = styled(Card)`
   background-color: ${uiColors.gray.light3};
@@ -30,7 +30,7 @@ const QuizHeader = styled('div')`
 
 const QuizSubtitle = styled('p')`
   font-weight: 500;
-  font-size: 14px;
+  font-size: ${theme.fontSize.small};
   line-height: 16.59px;
   color: ${uiColors.gray.base};
   margin: 0 !important;
@@ -66,12 +66,12 @@ const QuizCompleteSubtitle = ({ question }) => {
   );
 };
 
-const SubmitButton = ({ setHasSubmitted, selectedResponse, quizResponseObj }) => {
+const SubmitButton = ({ setIsSubmitted, selectedResponse, quizResponseObj }) => {
   const handleChoiceClick = useCallback(() => {
     if (selectedResponse) {
-      setHasSubmitted(true);
+      setIsSubmitted(true);
     }
-  }, [setHasSubmitted, selectedResponse]);
+  }, [setIsSubmitted, selectedResponse]);
 
   return (
     <StyledButton onClick={handleChoiceClick} variant="default">
@@ -92,7 +92,7 @@ const createQuizResponseObj = (question, quizId, selectedResponse, project) => {
 const QuizWidget = ({ nodeData: { children, options } }) => {
   const [question, ...choices] = children;
   const [selectedResponse, setSelectedResponse] = useState();
-  const [hasSubmitted, setHasSubmitted] = useState(false);
+  const [isSubmitted, setIsSubmitted] = useState(false);
   const quizId = options?.['quiz-id'];
   const { project } = useSiteMetadata();
   return (
@@ -107,12 +107,12 @@ const QuizWidget = ({ nodeData: { children, options } }) => {
             idx={i}
             selectedResponseIdx={selectedResponse?.index}
             setSelectedResponse={setSelectedResponse}
-            hasSubmitted={hasSubmitted}
+            isSubmitted={isSubmitted}
           />
         ))}
-        {!hasSubmitted && (
+        {!isSubmitted && (
           <SubmitButton
-            setHasSubmitted={setHasSubmitted}
+            setIsSubmitted={setIsSubmitted}
             selectedResponse={selectedResponse}
             quizResponseObj={createQuizResponseObj(question, quizId, selectedResponse, project)}
           />
